refining flatpickr plugin
